# PythonApi/openapi.json
{"openapi": "3.1.0", "info": {"title": "Movies API", "description": "Movies API with Todo Classification", "version": "v1"}, "paths": {"/": {"get": {"summary": "Redirect To Swagger", "operationId": "redirect_to_swagger__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/todos/classify": {"post": {"summary": "Classify Todo", "description": "Classify a todo item description into one of the categories: Work, Home, or Other.\nUses OpenAI API for classification.", "operationId": "classify_todo_api_todos_classify_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TodoClassificationRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TodoClassificationResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "TodoCategory": {"type": "string", "enum": ["Work", "Home", "Other"], "title": "TodoCategory"}, "TodoClassificationRequest": {"properties": {"description": {"type": "string", "title": "Description"}}, "type": "object", "required": ["description"], "title": "TodoClassificationRequest"}, "TodoClassificationResponse": {"properties": {"description": {"type": "string", "title": "Description"}, "category": {"$ref": "#/components/schemas/TodoCategory"}, "confidence": {"type": "number", "title": "Confidence"}}, "type": "object", "required": ["description", "category", "confidence"], "title": "TodoClassificationResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}